+++ Sources/Modules/NeuralNetwork/neuralNetwork.h
@@ -36,7 +36,7 @@ static OutputLayer * xConstructOutputLayer (Matrix * outputLayerMatrix, Matrix *
 
 NerualNetwork * xConstructNeuralNetwork (Matrix * inputMatrix, uint16_t numHiddenLayers, Matrix * hiddenLayerMatricies[numHiddenLayers], Matrix * hiddenLayerBiases[numHiddenLayers], Matrix * outputLayerMatrix, Matrix * outputLayerBiases);
 Matrix * xForwardPropagation (NerualNetwork * NN, e_FunctionOption activationFunction);
-void vTrainNeuralNetwork (NerualNetwork * NN, Matrix * expectedOutputMatrix, uint8_t numEpochs, double learningRate);
+void vTrainNeuralNetwork (NerualNetwork * NN, Matrix * expectedOutputMatrix, uint16_t numEpochs, double learningRate);
 
 void vPrintAllLayers (NerualNetwork * NN);
 static void vApplyActivationFunction(Matrix * m, ActivationFunction * function);
